chore: address PR feedback
move classes to separate folder, comment cleanup, small refactors
