+++ Aplib.Core/Intent/Tactics/PrimitiveTactic.cs
@@ -1,5 +1,6 @@
 ﻿using Aplib.Core.Belief.BeliefSets;
 using Aplib.Core.Intent.Actions;
+using Aplib.Core.Logging;
 using System.Collections.Generic;
 
 namespace Aplib.Core.Intent.Tactics
@@ -85,7 +86,7 @@ namespace Aplib.Core.Intent.Tactics
             => IsActionable(beliefSet) ? _action : null;
 
         /// <inheritdoc/>
-        public override IEnumerable<ILoggable> GetChildren() =>
+        public override IEnumerable<ILoggable> GetLogChildren() =>
             _action is ILoggable loggable ? new[] { loggable } : System.Linq.Enumerable.Empty<ILoggable>();
     }
 }
